+++ CHANGELOG.md
@@ -11,6 +11,11 @@
 - [#99](https://github.com/clojure-emacs/clojure-ts-mode/pull/99): Fix bug in `clojure-ts-align` when nested form has extra spaces.
 - [#99](https://github.com/clojure-emacs/clojure-ts-mode/pull/99): Fix bug in `clojure-ts-unwind` when there is only one expression after
   threading symbol.
+- Introduce `clojure-ts-jank-use-cpp-parser` customization which allows
+  highlighting C++ syntax in Jank `native/raw` forms.
+- Introduce `clojure-ts-clojurescript-use-js-parser` customization which allows
+  highlighting JS syntax in ClojureScript `js*` forms.
+
 
 ## 0.4.0 (2025-05-15)